Set up Ansible to provision git on build
This is a major commit that introduces many broad changes. The end
product is that the control machine is able to provision git on
the build machine using Ansible, and should be able to provision
any other program. However, for now, the command to provision other
machines from the control server must be issued manually:

    sudo ansible-playbook playbook.yml -i inventory.yml

This is for several reasons:
- Vagrant is configured to setup the control machine first, so it
  cannot provision the other machines at create time
- The inventory file is not recognised by the automatic provisioner
  (i.e. vagrant provision CONTROL_MACHINE) and so must be specified
  manually
- Provisioning must be carried out as sudo

The first and second issue should be relatively simple to deal with
by appropriately configuring Vagrant, while the third should be
manageable by appropriately configuring Ansible.

The hosts config file has been changed to a yaml-format Ansible
inventory, although it remains a source of truth for network to the
Vagrantfile for network configuration.

An issue was encountered in which machines could not ssh into one
another, with an error of:

    Permission denied (publickey,gssapi-keyex,gssapi-with-mic)

This was bypassed with a [hacky solution](https://devops.stackexchange.com/questions/1237/how-do-i-configure-ssh-keys-in-a-vagrant-multi-machine-setup)
which involves the injection of an ssh key pair each machine.
This key pair must be generated in the project directory with
`ssh-keygen` before setting up any machine.
Since ansible must run as root, these keys must (also) be copied
to the home of the root user.

Note that the ip adresses defined in the inventory should be wiped
on a future commit to prevent network clashes if this project is
used on multiple machines.