Provide manual build trigger instructions
True automatic building and deployment requires the build server
to be visible to GitHub (i.e. it must host a live HTTP server
accessible over the internet), which is obviously not realistic
in a cheap development scenario. Instead, builds can be triggered
from within the network by accessing a specified URL.